chore(refactor): extract /api/spaces/* into routes/spaces.ts; delete spaces-routes.ts
Third bucket of the index.ts route extraction. Also resolves the
pre-existing inconsistency the audit flagged: spaces handling was split
across two files using two different body-reading styles.

Routes consolidated into server/src/routes/spaces.ts (8 handlers):
  GET    /api/spaces
  POST   /api/spaces/from-folder       (was in spaces-routes.ts)
  POST   /api/spaces/from-path
  GET    /api/spaces/:id/sources
  POST   /api/spaces/:id/sources
  DELETE /api/spaces/:id/sources/:source_id
  PATCH  /api/spaces/:id               (was in spaces-routes.ts)
  DELETE /api/spaces/:id               (was in spaces-routes.ts)

Cleanup:
- server/src/spaces-routes.ts deleted (legacy req.on("data") body style
  superseded by readJsonBody / sendJson via RouteCtx).
- handleSpacesRequest call removed from index.ts.
- All eight handlers use safeDecode for path-param decoding (was
  unprotected in 4 of them, would have crashed on malformed URL encoding).

Net across the repo: index.ts is now 1481 lines — down 737 lines (-33%)
from the pre-audit baseline of 2218.

No behavioural change beyond the malformed-URL guard (was: server
process crash; now: 400 response).
